Tweak SB queue deinitialization logic
This commit makes a couple changes to destructor/deinit-related logic to
enable programs to interact with SB queues polymorphically:

1) This commit marks the SB_Base destructor virtual, which is necessary to
ensure that the destructors of derived classes get called when working
with SB_Base pointers that may point to an instance of a derived class.

2) This commit updates the SB_pcie destructor to be responsible for
deinitializing the device and host, and gets rid of the destructors in
SBRX_pcie and SBTX_pcie.

An alternative would be making the SB_pcie destructor virtual, but this
results in redundant calls to deinit_host(). Since it doesn't seem
dangerous to run deinit_dev() in the TLM case, I figured it would be
cleanest to just make this logic part of SB_pcie.