Update UMI submodule
This version of UMI no longer includes Switchboard as a submodule, which
prevents multiple copies of SB from being part of the tree if you
perform a recursive submodule update.

This is the newest commit of UMI that does not include any module
interface changes from the previously used commit.